fix: resets Sort filter on clearing of searchstring
This commit adds a new useSelector hook in the Filter component to retrieve the 'data' state from the Redux store. This will be used for filtering options.

The useEffect hook is also updated to reset the filter and select the "hot" option when the search string is empty.

---

chore: Import selectFilter function in Search component

This commit imports the selectFilter function from contentSlice in order to use it later for dispatching actions related to selecting filters.